requirements.txt
alembic==1.0.0
aniso8601==3.0.2
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
pytz==2018.5
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.11
urllib3==1.23
Werkzeug==0.14.1
